RF+ENH(TST): smoke test being able to export search in json and yaml
