+++ datalad/metadata/tests/test_search.py
@@ -19,6 +19,7 @@ from datalad.tests.utils import assert_in
 from datalad.tests.utils import assert_is_generator
 from datalad.tests.utils import with_tempfile
 from datalad.tests.utils import with_testsui
+from datalad.tests.utils import SkipTest
 from datalad.support.exceptions import NoDatasetArgumentFound
 
 from os.path import join as opj
@@ -118,4 +119,33 @@ def _check_mocked_install(central_dspath, mock_install):
         list(gen), [(loc, report)
                     for loc, report in _mocked_search_results])
     mock_install.assert_called_once_with(central_dspath, source='///')
+
+
+@with_tempfile
+def test_our_metadataset_search(tdir):
+    # smoke test for basic search operations on our super-megadataset
+    # expensive operation but ok
+    ds = install(path=tdir, source="///")
+    assert list(ds.search('.', report='*', regex=True))
+    assert list(ds.search('.', report='*'))
+    assert list(ds.search('.', report_matched=True))
+
+    # and either we could provide output in different formats
+    import simplejson
+    from datalad.utils import swallow_outputs
+    from datalad.api import search_
+    with swallow_outputs() as cmo:
+        assert list(search_('.', report='*', regex=True, format='json', dataset=ds))
+        out = cmo.out
+    # since this one is just absorbs all first, we can't go one by one
+    assert simplejson.loads(out)
+
+    try:
+        import yaml
+    except ImportError:
+        raise SkipTest("no yaml module")
+    with swallow_outputs() as cmo:
+        assert list(search_('.', report='*', regex=True, format='yaml', dataset=ds))
+        out = cmo.out
+    assert yaml.load(out)
